Switch UFlow class from wrapping prefect Flow object to extending it
This is a more natural fit for the intended behavior as it allows us to use the parent class methods while adding additional functionality such as wrapping a flow run with mlflow. It also removes the need for referencing the underlying flow object with UFlow.flow, which is confusing. This also eliminates the need of the __enter__ and __exit__ methods.