Add CLI argument that allows setting custom name of pipeline object
The user may choose to set the name of the flow object to anything they want when defining the flow. The converter should be able to accomodate this and be flexible. This functionality is now provided as a CLI optional argument (default value of the name of the flow object is "flow").